+++ src/android/org/linphone/chat/GroupChatFragment.java
@@ -286,10 +286,10 @@ public class GroupChatFragment extends Fragment implements ChatRoomListener, Con
 			EventLog eventLog = (EventLog)obj;
 			if (eventLog.getType() == EventLog.Type.ConferenceChatMessage) {
 				ChatMessage message = eventLog.getChatMessage();
-				if (message.getAppdata() != null) {
+				if (message.getAppdata() != null && !message.isOutgoing()) {
 					File file = new File(message.getAppdata());
 					if (file.exists()) {
-						file.delete();
+						file.delete(); // Delete downloaded file from incoming message that will be deleted
 					}
 				}
 			}
